Fixed: bugs in handle array Fixed: bugs in hash map Added: material render data Changed sizeof render handles (2^32 # of handles must be enough) Finished: obj importer

+++ src/renderer/resource_managers/model.h
@@ -9,7 +9,7 @@
 namespace Veng
 {
 
-enum class meshRenderHandle : u64 {};
+enum class meshRenderHandle : u32 {};
 const meshRenderHandle INVALID_MESH_RENDER_HANDLE = (meshRenderHandle)0;
